+++ etc/starter-kitty-validators.api.md
@@ -6,7 +6,6 @@
 
 /// <reference types="node" />
 
-import { URL as URL_2 } from 'url';
 import { z } from 'zod/v4';
 
 // @public
@@ -66,28 +65,6 @@ export interface UrlValidatorOptions {
     baseOrigin?: string;
     // Warning: (ae-forgotten-export) The symbol "UrlValidatorWhitelist" needs to be exported by the entry point index.d.ts
     whitelist?: UrlValidatorWhitelist;
-}
-
-// @public
-export const webhookUrlSchema: z.ZodPipe<z.ZodURL, z.ZodTransform<URL_2, string>>;
-
-// @public
-export class WebhookUrlValidationError extends Error {
-    constructor(message: string);
-}
-
-// @public
-export class WebhookUrlValidator {
-    constructor(options?: WebhookUrlValidatorOptions);
-    fetch(url: string | URL, init?: RequestInit): Promise<Response>;
-    validate(url: string | URL): URL;
-    validateAsync(url: string | URL): Promise<URL>;
-}
-
-// @public
-export interface WebhookUrlValidatorOptions {
-    // Warning: (ae-forgotten-export) The symbol "DnsLookupFn" needs to be exported by the entry point index.d.ts
-    lookup?: DnsLookupFn;
 }
 
 ```
