+++ packages/validators/README.md
@@ -73,12 +73,14 @@ window.location.pathname = validator.parsePathname(redirectUrl, '/home') // fall
 
 `WebhookUrlValidator` is the **inverse** of `UrlValidator`: instead of allowlisting known-safe hosts for redirects within your own app, it **blocklists** private/internal network targets for arbitrary, user-supplied URLs that your server will make outbound requests to - the shape of the problem when a user registers a webhook destination. There are two places to use it: when the URL is first saved, and every time you actually send to it.
 
+It lives under the `/server/webhook-url` subpath (not `/webhook-url`) because it resolves DNS with `node:dns/promises` - server-only code that a browser or edge bundle can't include. Only import it from server-side code.
+
 ### 1. Validating the URL when it's saved as config
 
 Use `webhookUrlSchema` wherever a webhook URL is taken as input, so a request to save an obviously unsafe URL is rejected before it ever reaches storage:
 
 ```ts
-import { webhookUrlSchema } from '@opengovsg/starter-kitty-validators/webhook-url'
+import { webhookUrlSchema } from '@opengovsg/starter-kitty-validators/server/webhook-url'
 
 const saveWebhookConfigSchema = z.object({
   url: webhookUrlSchema,
@@ -102,7 +104,7 @@ Not using zod for this input? `new WebhookUrlValidator().validate(rawUrl)` does 
 Use `WebhookUrlValidator.fetch` to actually deliver, every time an event fires - not just once at registration. It re-runs the sync checks, resolves the hostname, validates every resolved IP (catching DNS rebinding: a hostname that resolved to a safe address when saved but an internal one now, or that has a mix of safe and unsafe A/AAAA records), and only then makes the request, rejecting redirects instead of following them:
 
 ```ts
-import { WebhookUrlValidator } from '@opengovsg/starter-kitty-validators/webhook-url'
+import { WebhookUrlValidator } from '@opengovsg/starter-kitty-validators/server/webhook-url'
 
 const webhookValidator = new WebhookUrlValidator()
