+++ packages/validators/src/index.ts
@@ -8,4 +8,3 @@ export type * from '@/common/errors'
 export * from '@/email'
 export * from '@/path'
 export * from '@/url'
-export * from '@/webhook-url'
